+++ FRW/Modules/Staking/StakingDetail/StakingDetailViewModel.swift
@@ -39,7 +39,7 @@ class StakingDetailViewModel: ObservableObject {
     var availableAmount: Double {
         let token = WalletManager.shared.flowToken
         let balance = WalletManager.shared.getBalance(byId: token?.contractId ?? "")
-        return balance
+        return balance.doubleValue
     }
 
     var currentProgress: Int {
